requirements.txt
requests==2.28.2
flask==2.2.2
pyinstaller==5.7.0
PyExecJS==1.5.1